Addresses event deprecation in 1.12
https://github.com/tastybento/ASkyBlock-Bugs-N-Features/issues/383
